feat: add parent component for examples
