+++ demo/index.jsx
@@ -1,5 +1,9 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
+import {HashRouter} from "react-router-dom";
+
 const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(<App />);
+root.render(<HashRouter>
+    <App/>
+</HashRouter>);
